機能別モジュール化の実装
- models.py: データモデル（LineType、Cardinality、Column、Table）
- rendering.py: エッジとIE記法レンダリング（Edge）
- canvas.py: キャンバス・ノード管理（Canvas、Node）
- er_diagram.py: ER図管理とレイアウト（ERDiagram）
- stencil.py: ステンシル機能（既存維持）

既存機能の互換性を保持し、責任分離とテスト性を向上
