+++ planning/planning/__init__.py
@@ -1,4 +1,5 @@
-from .diagram import ERDiagram
+from .er_diagram import ERDiagram
 from .stencil import TableStencil, Stencil
+from .models import Table, Column, LineType, Cardinality
 
-__all__ = ['ERDiagram', 'TableStencil', 'Stencil']
+__all__ = ['ERDiagram', 'TableStencil', 'Stencil', 'Table', 'Column', 'LineType', 'Cardinality']
